+++ cf/data/fragment/__init__.py
@@ -1,5 +1,3 @@
 from .fullfragmentarray import FullFragmentArray
-
-# from .missingfragmentarray import MissingFragmentArray
 from .netcdffragmentarray import NetCDFFragmentArray
 from .umfragmentarray import UMFragmentArray